requirements.txt
pandas
sqlalchemy
requests
boto3
beautifulsoup4
lxml
python-dotenv  
pymysql
selenium
webdriver-manager            
undetected-chromedriver      